services/emotion_emotieff/requirements.txt
fastapi
uvicorn[standard]
pydantic
emotiefflib
numpy
pillow
opencv-python-headless